% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artcatr.R
\name{print.artcatr}
\alias{print.artcatr}
\title{Print an \code{artcatr} object}
\usage{
\method{print}{artcatr}(x, digits = 3, probtable = TRUE, ...)
}
\arguments{
\item{x}{An object returned by artcatr().}

\item{digits}{Number of significant digits to round probabilities to.}

\item{probtable}{Logical; include the table of anticipated probabilities?}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an \code{artcatr} object
}
