% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artcatr.R
\name{print.summary.artcatr}
\alias{print.summary.artcatr}
\title{Print a \code{summary.artcatr} object}
\usage{
\method{print}{summary.artcatr}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=summary.artcatr]{summary.artcatr()}}.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a \code{summary.artcatr} object
}
